+++ .travis-bazelrc
@@ -12,3 +12,4 @@ build --verbose_failures
 build --sandbox_debug
 build --test_output=errors
 build --flaky_test_attempts=5
+build --features=race
