+++ shared/p2p/BUILD.bazel
@@ -52,17 +52,22 @@ go_test(
         "@com_github_libp2p_go_libp2p_peer//:go_default_library",
         "@com_github_libp2p_go_libp2p_swarm//testing:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
-        "@com_github_sirupsen_logrus//hooks/test:go_default_library",
     ],
 )
 
 # by default gazelle tries to add all the test files to the  first
-# go_test; we want to treat feed_concurrent_test.go differently
-# gazelle:exclude feed_concurrent_test.go
+# go_test; we need to exclude the tests that we want to configure
+# in a particular way
+# gazelle:exclude racy_test.go
 
 go_test(
-    name = "go_feed_concurrent_write_test",
-    srcs = ["feed_concurrent_test.go"],
+    name = "go_racy_test",
+    srcs = ["racy_test.go"],
     embed = [":go_default_library"],
-    race = "on",
+    race = "off",  # TODO(#377): fix issues with race detection testing.
+    deps = [
+        "@com_github_libp2p_go_libp2p//p2p/host/basic:go_default_library",
+        "@com_github_libp2p_go_libp2p_swarm//testing:go_default_library",
+        "@com_github_sirupsen_logrus//hooks/test:go_default_library",
+    ],
 )
