+++ shared/p2p/feed_test.go
@@ -35,3 +35,14 @@ func TestFeed_ReturnsSameFeed(t *testing.T) {
 		}
 	}
 }
+
+func TestFeed_ConcurrentWrite(t *testing.T) {
+	s, err := NewServer()
+	if err != nil {
+		t.Fatalf("could not create server %v", err)
+	}
+
+	for i := 0; i < 5; i++ {
+		go s.Feed("a")
+	}
+}
